requirements.txt
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
charset-normalizer==3.4.0
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pyproject_hooks==1.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
urllib3==2.2.3
virtualenv==20.27.1
djangorestframework==3.16.1
